requirements.txt
asgiref==3.8.1
Django==5.2
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
sqlparse==0.5.3
uritemplate==4.1.1
gunicorn
djangorestframework
djangorestframework-simplejwt
django-environ
drf-spectacular
django_redis
Pillow
django-ratelimit
pylint-django